ch05 / ex26 : `%ld`

+++ ch05/ch05/ex26.cpp
@@ -3,6 +3,6 @@
 int main() {
     int arr[10] = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
 
-    printf("arr의 값 : %d\n", arr);
-    printf("&arr[0]의 값 : %d\n", &arr[0]);
+    printf("arr의 값 : %ld\n", (long) arr);
+    printf("&arr[0]의 값 : %ld\n", (long) &arr[0]);
 }
